Refactored and cleaned up stl-example.
