+++ Examples/stl-example/STL-bin/Win32/Debug/ALL_BUILD/ALL_BUILD.log
@@ -1,2 +1,1 @@
-﻿  Building Custom Rule C:/Users/KSeam/Documents/Courses/UBCO/Year 4/COSC 499/Capstone3DModelling/Examples/stl-example/CMakeLists.txt
-  CMake does not need to re-run because C:/Users/KSeam/Documents/Courses/UBCO/Year 4/COSC 499/Capstone3DModelling/Examples/stl-example/STL-bin/CMakeFiles/generate.stamp is up-to-date.
+﻿
