+++ Examples/stl-example/STL-bin/Win32/Debug/ZERO_CHECK/ZERO_CHECK.log
@@ -1,2 +1,8 @@
 ﻿  Checking Build System
-  CMake does not need to re-run because C:/Users/KSeam/Documents/Courses/UBCO/Year 4/COSC 499/Capstone3DModelling/Examples/stl-example/STL-bin/CMakeFiles/generate.stamp is up-to-date.
+  CMake is re-running because C:/Users/KSeam/Documents/Courses/UBCO/Year 4/COSC 499/Capstone3DModelling/Examples/stl-example/STL-bin/CMakeFiles/generate.stamp is out-of-date.
+    the file 'C:/Users/KSeam/Documents/Courses/UBCO/Year 4/COSC 499/Capstone3DModelling/Examples/stl-example/CMakeLists.txt'
+    is newer than 'C:/Users/KSeam/Documents/Courses/UBCO/Year 4/COSC 499/Capstone3DModelling/Examples/stl-example/STL-bin/CMakeFiles/generate.stamp.depend'
+    result='-1'
+  -- Configuring done
+  -- Generating done
+  -- Build files have been written to: C:/Users/KSeam/Documents/Courses/UBCO/Year 4/COSC 499/Capstone3DModelling/Examples/stl-example/STL-bin
