+++ Examples/stl-example/VTK.h
@@ -1,0 +1,3 @@
+#pragma once
+std::string utf8_encode(const std::wstring &wstr);
+int VTKmain(LPWSTR filePathReferance, LPWSTR filePathProduction);
