+++ Examples/stl-example/fileDialogue.h
@@ -1,0 +1,2 @@
+#pragma once
+LPWSTR fileDialogue();
